[EDIT] small adjustment

+++ Classes/Tools/RedirectLogin.php
@@ -5,6 +5,7 @@ namespace RKW\RkwRegistration\Tools;
 use \RKW\RkwBasics\Helper\Common;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use RKW\RkwBasics\Service\CookieService;
+use RKW\RkwBasics\Service\CacheService;
 
 /*
  * This file is part of the TYPO3 CMS project.
@@ -104,6 +105,7 @@ class RedirectLogin implements \TYPO3\CMS\Core\SingletonInterface
                     && ($checkedUrl = $this->checkRedirectUrl($referrer))
                 ) {
                     $feAuth->setKey('ses', 'rkw_registration_redirect_referrer', $checkedUrl);
+                    // add also to rkw cookie
                     CookieService::setKey('rkw_registration_redirect_referrer', $checkedUrl);
                     $this->getLogger()->log(\TYPO3\CMS\Core\Log\LogLevel::DEBUG, sprintf('Referrer redirect set to %s.', $checkedUrl));
                 }
@@ -120,7 +122,10 @@ class RedirectLogin implements \TYPO3\CMS\Core\SingletonInterface
                     && ($checkedUrl = $this->checkRedirectUrl($xdlUrl))
                 ) {
                     $feAuth->setKey('ses', 'rkw_registration_redirect_xdl_url', $checkedUrl);
-                    CookieService::setKey('rkw_registration_redirect_xdl_url', $checkedUrl);
+                    if ($GLOBALS['TYPO3_CONF_VARS']['FE']['cookieNameRkwBasics']) {
+                        // add also to rkw cookie
+                        CookieService::setKey('rkw_registration_redirect_xdl_url', $checkedUrl);
+                    }
                     $this->getLogger()->log(\TYPO3\CMS\Core\Log\LogLevel::DEBUG, sprintf('XDL redirect set to %s.', $checkedUrl));
                 }
             }
@@ -218,7 +223,11 @@ class RedirectLogin implements \TYPO3\CMS\Core\SingletonInterface
 
             // reset referrer in cookie. Everything else is kept - except for logout!
             $feAuth->setKey('ses', 'rkw_registration_redirect_referrer', null);
-            CookieService::removeKey('rkw_registration_redirect_referrer');
+
+            if ($GLOBALS['TYPO3_CONF_VARS']['FE']['cookieNameRkwBasics']) {
+                // remove it also from RkwCookie
+                CookieService::removeKey('rkw_registration_redirect_referrer');
+            }
 
         }
 
@@ -296,8 +305,13 @@ class RedirectLogin implements \TYPO3\CMS\Core\SingletonInterface
             // reset everything in cookie
             $feAuth->setKey('ses', 'rkw_registration_redirect_referrer', null);
             $feAuth->setKey('ses', 'rkw_registration_redirect_xdl_url', null);
-            CookieService::removeKey('rkw_registration_redirect_referrer');
-            CookieService::removeKey('rkw_registration_redirect_xdl_url');
+
+
+            if ($GLOBALS['TYPO3_CONF_VARS']['FE']['cookieNameRkwBasics']) {
+                // remove it also from RkwCookie
+                CookieService::removeKey('rkw_registration_redirect_referrer');
+                CookieService::removeKey('rkw_registration_redirect_xdl_url');
+            }
         }
 
         return $url;
